Required Responders for GO decision Single Arm

Function for calculating the minimum required number of responders in the experimental group to make a GO decision in Settings 1 and 2.

req_resp(
  N_e,
  delta,
  confidence,
  e_a = 0.5,
  e_b = 0.5,
  h_a = 0.5,
  h_b = 0.5,
  RR_h = NULL,
  N_h = NULL,
  hist_RR_c = NULL,
  adapt = 1
)

Arguments

N_e
Sample Size in the experimental group.

delta
Required superiority to make a "GO" decision. Corresponds to \(\delta\).

confidence
Required confidence to make "GO" decision. Corresponds to \(\gamma\).

e_a
Alpha parameter of Beta Prior Distribution for the experimental response rate. Corresponds to \(\alpha_e\). Default is \(\frac{1}{2}\).

e_b
Beta parameter of Beta Prior Distribution for the experimental response rate. Corresponds to \(\beta_e\). Default is \(\frac{1}{2}\).

h_a
Alpha parameter of Beta Prior Distribution for the historical control response rate. Corresponds to \(\alpha_h\). Only needs to be specified, if RR_h and N_h are also specified. Default is \(\frac{1}{2}\).

h_b
Beta parameter of Beta Prior Distribution for the historical control response rate. Corresponds to \(\beta_h\). Only needs to be specified, if RR_h and N_h are also specified. Default is \(\frac{1}{2}\).

RR_h
Historical control response rate. Corresponds to \(p_h\). If specified together with N_h, function will use setting 2 from pdf.

N_h
Historical control sample size. Corresponds to \(n_h\). If specified together with RR_h, function will use setting 2 from pdf.

hist_RR_c
Point estimate of historical control repsonse rate. Corresponds to \(\hat{p_h}\). If specified, while RR_h and N_h are not specified, function will use setting 1 from pdf.

adapt
Level of adapting of experimental control rate to account for patient selection bias from phase II to phase III. Corresponds to \(\xi\). Default is 1, so no adapting.

Value

Integer.

Examples

# Setting 1
req_resp(
  N_e = 50, delta = 0.08,
  confidence = 0.6, hist_RR_c = 0.5
)

#> [1] 30

# Setting 2
req_resp(
  N_e = 50, delta = 0.08,
  confidence = 0.6, RR_h = 0.5, N_h = 50
)

#> [1] 31